# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 11.15.0
  description: |-
    (Sovereign) Agent Governed Experience — network REST API for memory
    submission, validation, and query.

    This machine-readable contract intentionally excludes the same-machine
    human CEREBRUM control plane under `/v1/dashboard/**`. Those routes are
    loopback-only management APIs, not remotely callable network APIs. Their
    authoritative contract is the “App-v23 CEREBRUM access-control routes”
    section of `docs/reference/rest-api.md`.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.
        Current clients should also send **X-Nonce**. After app-v17 activation,
        consensus binds delegated REST proofs to the exact canonical request,
        enforces the timestamp against block time, and consumes each proof once.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
  parameters:
    memoryId:
      name: memory_id
      in: path
      required: true
      description: Unique memory object identifier (UUID v7).
      schema:
        type: string
        format: uuid

  # -----------------------------------------------------------------------
  # Schemas
  # -----------------------------------------------------------------------
  schemas:
...
    MemoryType:
      type: string
      enum:
        - fact
        - observation
        - inference
        - task
      description: Classification of the memory object.
    MemoryStatus:
      type: string
      enum:
        - proposed
        - validated
        - committed
        - challenged
        - deprecated
      description: Lifecycle state of a memory object.
    VoteDecisionEnum:
      type: string
      enum:
        - accept
        - reject
        - abstain

    # -- Core domain objects -----------------------------------------------
    KnowledgeTriple:
      type: object
      required:
        - subject
        - predicate
        - object
      properties:
        subject:
          type: string
          description: Subject entity of the triple.
        predicate:
          type: string
          description: Relationship between subject and object.
        object:
          type: string
          description: Object entity of the triple.
    MemoryRecord:
      type: object
      required:
        - memory_id
        - submitting_agent
        - content
        - content_hash
        - memory_type
        - domain_tag
        - confidence_score
        - status
        - created_at
      properties:
        memory_id:
          type: string
          format: uuid
          description: Unique identifier (UUID v7).
        submitting_agent:
          type: string
          description: Hex-encoded Ed25519 public key of the proposing agent.
        content:
          type: string
          description: Natural-language memory content.
        content_hash:
          type: string
          description: SHA-256 hash of content (hex-encoded).
        embedding:
          type: array
          items:
            type: number
            format: float
          description: 1536-dimensional embedding vector.
        embedding_hash:
          type: string
          description: SHA-256 hash of the embedding vector (hex-encoded).
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        domain_tag:
          type: string
          description: Domain or topic tag for the memory.
        confidence_score:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: >-
            Current confidence at query time: the agent-assigned score after
            time decay and the corroboration boost are applied.
        initial_confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: >-
            Stored (undecayed) confidence — the on-chain value set at submission
            (corroboration never rewrites it), before time decay. Sibling to the
            decayed confidence_score. Present for local memories; omitted (absent)
            for federated results, where only the serving peer's already-decayed
            value is available. Not guaranteed on every result.
        corroboration_count:
          type: integer
          minimum: 0
          description: >-
            Number of distinct corroborating agent IDs backing this memory — the
            multiplier behind the corroboration boost in confidence_score.
            Returned on recall paths (/v1/memory/query, /search, /hybrid) and
            the detail endpoint so readers can tell a low score caused by no
            corroboration (a fresh, untested belief) from one caused by time
            decay (a once-solid fact).
        challenge_count:
          type: integer
          minimum: 0
          description: >-
            Number of distinct challenger IDs in the off-chain lifetime audit
            projection. This is evidence history, not an open-vote count.
        evidence_counts_available:
          type: boolean
          description: >-
            True only when both count queries succeeded and no durable
            recovery/repair-incomplete marker was detected. It is not an
            attestation against out-of-band partial SQL audit-table corruption.
            When false, corroboration_count and challenge_count may still be
            reconstructed canonical lower bounds after pristine state sync or
            repair; zero is not proof that no historical evidence existed.
        challenge_round:
          type: integer
          format: int64
          minimum: 1
          description: Current app-v21 challenge round; absent when no weighted round is open.
        current_challenger_count:
          type: integer
          format: int32
          minimum: 1
          description: Distinct challengers accrued in the current open app-v21 round.
        required_challengers:
          type: integer
          format: int32
          minimum: 1
          description: Distinct challengers required to deprecate in the current app-v21 round.
        disputed:
          type: boolean
          description: >-
            True when the memory remains recallable while challenged. When true,
            confidence_score already includes the disputed confidence haircut.
        classification:
          type: integer
          minimum: 0
          maximum: 4
          description: >-
            Per-record clearance level: 0=PUBLIC, 1=INTERNAL, 2=CONFIDENTIAL,
            3=SECRET, 4=TOPSECRET.
        status:
          $ref: "#/components/schemas/MemoryStatus"
        parent_hash:
          type: string
          description: Hash of the parent memory (for lineage), if any.
        task_status:
          type: string
          description: >-
            For task-type memories: planned, in_progress, done, or dropped.
        knowledge_triples:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeTriple"
          description: Extracted knowledge graph triples.
        votes:
          type: array
          items:
            type: object
            properties:
              voter_id:
                type: string
              decision:
                $ref: "#/components/schemas/VoteDecisionEnum"
              rationale:
                type: string
              voted_at:
                type: string
                format: date-time
          description: Votes cast on this memory.
        corroborations:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: string
              evidence:
                type: string
              created_at:
                type: string
                format: date-time
          description: Corroborations supporting this memory.
        linked_memories:
          type: array
          items:
            type: object
          description: Memories linked to this one.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the memory was proposed.
        committed_at:
          type: string
          format: date-time
          description: Timestamp when the memory reached committed status.

    # -- Request bodies ----------------------------------------------------
...
    FilterInfo:
      type: object
      description: >-
        Identifies authorization filters that were applied. Before app-v23 a
        server may also emit legacy inventory counts. App-v23 omits denied/raw
        counts, and presence of this object is not proof that a row was hidden.
      properties:
        by:
          type: array
          items:
            type: string
          description: >-
            Filter identifiers applied. Values: rbac_submitting_agents,
            classification.
        total_before_filter:
          type: integer
          deprecated: true
          description: Legacy/pre-app-v23 only; omitted by app-v23.
        visible:
          type: integer
          deprecated: true
          description: Legacy/pre-app-v23 only; omitted by app-v23.
        hidden_count:
          type: integer
          deprecated: true
          description: Legacy/pre-app-v23 only; omitted by app-v23.
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        reason_code:
          type: string
          description: Stable sanitized machine-readable denial reason, when available.
        remedy:
          type: string
          description: Exact operator action that can make a later request valid.
        retryable:
          type: boolean
          description: Whether automatically repeating the same request is safe.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: >-
        Caller authenticated but lacks authorisation, or the request was
        rejected during on-chain finalisation (e.g. quorum-gated tx without
        an executed proposal). The `detail` field carries the upstream
        rejection reason.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
paths:
...
  /v1/memory/list:
    get:
      operationId: listMemories
      summary: Paginated memory list with RBAC agent isolation.
      description: >-
        App-v23 applies live authorization before visible offset/limit. When
        has_more is true, total is an authorization-safe visible lower bound
        and total_exact is false. Once the visible stream is exhausted,
        total_exact is true and total is exact. Denied/raw counts are never
        returned. Visible offset is capped at 7,900 and one request examines at
        most 8,192 raw authorization candidates. A broader request returns 422
        Query too broad and must be narrowed.
      tags:
        - Memory
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 7900
          description: App-v23 maximum is 7,900 visible records.
        - name: domain
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: agent
          in: query
          schema:
            type: string
          description: Filter by submitting agent ID.
      responses:
        "200":
          description: Paginated list of memories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      $ref: "#/components/schemas/MemoryRecord"
                  total:
                    type: integer
                    description: >-
                      Visible total or authorization-safe lower bound; inspect
                      total_exact.
                  limit:
                    type: integer
                  offset:
                    type: integer
                  has_more:
                    type: boolean
                    description: Whether another currently visible row exists.
                  total_exact:
                    type: boolean
                    description: Whether total is the exact visible total.
                  filtered:
                    $ref: "#/components/schemas/FilterInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: >-
            Offset exceeds 7,900 or app-v23 authorization would require
            examining more than 8,192 raw candidates. Narrow filters or page
            sequentially.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Current app-v23 record authorization state is unavailable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/memory/{memory_id}:
    get:
      operationId: getMemory
      summary: Retrieve a single memory by ID.
      description: >-
        App-v23 applies the same live enrollment, legacy-envelope,
        domain/group/grant, and clearance decision as recall/list. Authorship
        is provenance, not a read bypass.
      tags:
        - Memory
      parameters:
        - $ref: "#/components/parameters/memoryId"
      responses:
        "200":
          description: Full memory record with votes and corroborations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Current app-v23 record authorization state is unavailable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          $ref: "#/components/responses/InternalError"
...